* subversion/libsvn_fs_fs/caching.c:   (manifest_deserialize): Remove an unneeded cast added in r1040115.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1040126 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_fs_fs/caching.c
@@ -135,7 +135,7 @@ manifest_deserialize(void **out,
 {
   apr_array_header_t *manifest = apr_array_make(pool,
                                        (int) (data_len / sizeof(apr_off_t)),
-                                       (int) sizeof(apr_off_t));
+                                       sizeof(apr_off_t));
   memcpy(manifest->elts, data, data_len);
   manifest->nelts = (int) (data_len / sizeof(apr_off_t));
   *out = manifest;
